+++ src/core/common/util.hpp
@@ -48,6 +48,7 @@ static_assert(false, "C/C++ compiler not supported");
 #endif
 
 namespace vb {
+
 ///
 /// @brief Count leading zeros implementation
 ///
@@ -546,6 +547,32 @@ template <class Dest> inline Dest readNextValue(uint8_t const **const ptr) VB_NO
   *ptr = pSubI(*ptr, sizeof(Dest));
   return readFromPtr<Dest>(*ptr);
 }
+
+///
+/// @brief Represents a constant
+///
+union ConstUnion {
+  uint32_t u32; ///< 32-bit integer
+  uint64_t u64; ///< 64-bit integer
+  float f32;    ///< 32-bit float
+  double f64;   ///< 64-bit float
+
+  ///
+  /// @brief Get the raw, reinterpreted value of the float as an integer
+  ///
+  /// @return uint32_t Raw, reinterpreted value of the float
+  inline uint32_t rawF32() const VB_NOEXCEPT {
+    return bit_cast<uint32_t>(f32);
+  }
+
+  ///
+  /// @brief Get the raw, reinterpreted value of the float as an integer
+  ///
+  /// @return uint64_t Raw, reinterpreted value of the float
+  inline uint64_t rawF64() const VB_NOEXCEPT {
+    return bit_cast<uint64_t>(f64);
+  }
+};
 
 } // namespace vb
 #endif