+++ src/core/compiler/backend/x86_64/x86_64_backend.cpp
@@ -631,8 +631,8 @@ void Backend::iterateScratchRegsAndGlobals(FunctionRef<void(StackElement const &
     lambda(StackElement::scratchReg(WasmABI::fpr[regPos], StackType::SANULL));
   }
 
-  for (uint32_t globalIdx{0U}; globalIdx < moduleInfo_.numNonImportedGlobals; globalIdx++) {
-    lambda(StackElement::global(globalIdx));
+  for (uint32_t i{0U}; i < moduleInfo_.numNonImportedGlobals; i++) {
+    lambda(StackElement::nonImportGlobal(i, moduleInfo_.numImportedGlobals));
   }
 }